flask_opentelemetry/requirements.txt
flask 
flask_sqlalchemy
sqlalchemy.orm
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
opentelemetry-distro
python-dotenv